@@ -152,10 +152,26 @@ namespace Nager.Date.UnitTest.Common
         [TestMethod]
         public void CheckPublicHolidayWithDateFilter1()
         {
-            var items = DateSystem.GetPublicHoliday(new DateTime(2016, 5, 1), new DateTime(2018, 5, 31), CountryCode.DE);
+            this.CheckPublicHolidayWithDateFilter1(new DateTime(2016, 5, 1), new DateTime(2018, 5, 31));
+            this.CheckPublicHolidayWithDateFilter1(new DateTime(2016, 5, 1, 0, 0, 1), new DateTime(2018, 5, 31, 23, 59, 59));
+            this.CheckPublicHolidayWithDateFilter1(new DateTime(2016, 5, 1, 12, 30, 0), new DateTime(2018, 5, 31, 0, 0, 0));
+            this.CheckPublicHolidayWithDateFilter1(new DateTime(2016, 5, 1, 23, 59, 59), new DateTime(2018, 5, 31, 23, 59, 59));
+        }
 
+        private void CheckPublicHolidayWithDateFilter1(DateTime startDate, DateTime endDate)
+        {
+            var items = DateSystem.GetPublicHoliday(startDate, endDate, CountryCode.DE);
+
+            Assert.AreEqual(38, items.Count());
             Assert.IsTrue(items.First().Date > new DateTime(2016, 4, 28));
             Assert.IsTrue(items.Last().Date < new DateTime(2018, 6, 1));
+        }
+
+        [TestMethod]
+        [ExpectedException(typeof(ArgumentException), "startDate is before endDate")]
+        public void CheckPublicHolidayWithDateFilter2()
+        {
+            DateSystem.GetPublicHoliday(new DateTime(2016, 1, 2), new DateTime(2016, 1, 1), CountryCode.DE).First();
         }
 
         [TestMethod]
@@ -175,13 +191,6 @@ namespace Nager.Date.UnitTest.Common
         {
             var isPublicHoliday = DateSystem.IsOfficialPublicHolidayByCounty(new DateTime(2019, 8, 5), CountryCode.AU, "AU-NT");
             Assert.IsTrue(isPublicHoliday);
-        }
-
-        [TestMethod]
-        [ExpectedException(typeof(ArgumentException), "startDate is before endDate")]
-        public void CheckPublicHolidayWithDateFilter2()
-        {
-            var items = DateSystem.GetPublicHoliday(new DateTime(2016, 1, 2), new DateTime(2016, 1, 1), CountryCode.DE).First();
         }
     }
 }
