@@ -14,7 +14,7 @@ namespace Nager.Date.UnitTest.Country
             for (var year = DateTime.Now.Year; year < 3000; year++)
             {
                 var publicHolidays = DateSystem.GetPublicHoliday(year, CountryCode.ES);
-                Assert.AreEqual(35, publicHolidays.Count());
+                Assert.AreEqual(34, publicHolidays.Count());
             }
         }
 
