@@ -1,6 +1,6 @@
 ﻿@{
-    ViewData["Title"] = "Api";
-    ViewData["Description"] = "A free API for public holiday data support currently over 90 countries";
+    ViewData["Title"] = "Worldwide Public Holiday Api";
+    ViewData["Description"] = "A free API for public holiday data. We currently support over 100 countries.";
 }
 
 <div class="container">
